feat: non_db_request_variant macro
