+++ mysk-lib-macros/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod fetch_level_variant;
 pub mod traits;
